refactor(codegen): rename top-level registry to generator-registry
