+++ scripts/codegen/src/generator-registry.ts
