+++ scripts/codegen/src/generators/gen-contract.ts
@@ -2,11 +2,11 @@ import { mkdir, readdir, readFile, writeFile } from "node:fs/promises";
 import { dirname, resolve } from "node:path";
 import { fileURLToPath } from "node:url";
 import { parse as parseYaml } from "yaml";
+import type { GeneratorContext, GeneratorReport } from "../generator-registry.ts";
+import { registerGenerator } from "../generator-registry.ts";
 import { type Breakpoint, loadBreakpoints } from "../lib/breakpoints.ts";
 import { type FlatProp, type FlatSpec, flattenSpec } from "../lib/flatten.ts";
 import { pascalCase } from "../lib/pascal-case.ts";
-import type { GeneratorContext, GeneratorReport } from "../registry.ts";
-import { registerGenerator } from "../registry.ts";
 import { Spec as SpecSchema } from "../schema.ts";
 import { renderContract } from "./gen-contract/per-spec.ts";
 import { renderBarrel } from "./gen-contract/workspace/barrel.ts";
