+++ scripts/codegen/src/generators/gen-tests.ts
@@ -2,10 +2,10 @@ import { mkdir, readdir, readFile, writeFile } from "node:fs/promises";
 import { dirname, resolve } from "node:path";
 import { fileURLToPath } from "node:url";
 import { parse as parseYaml } from "yaml";
+import type { GeneratorContext, GeneratorReport } from "../generator-registry.ts";
+import { registerGenerator } from "../generator-registry.ts";
 import { flattenSpec } from "../lib/flatten.ts";
 import { pascalCase } from "../lib/pascal-case.ts";
-import type { GeneratorContext, GeneratorReport } from "../registry.ts";
-import { registerGenerator } from "../registry.ts";
 import { Spec as SpecSchema } from "../schema.ts";
 import type { Spec } from "./gen-contract.ts";
 import { renderReactFixtureFile } from "./gen-tests/per-spec/react-fixture.ts";
