+++ scripts/codegen/src/generators/gen-vue.ts
@@ -4,12 +4,12 @@ import { fileURLToPath } from "node:url";
 import { parse as parseYaml } from "yaml";
 import type { SpecAnalysis } from "../core/analysis.ts";
 import { computeAnalysis } from "../core/orchestrator.ts";
+import type { GeneratorContext, GeneratorReport } from "../generator-registry.ts";
+import { registerGenerator } from "../generator-registry.ts";
 import { type Breakpoint, loadBreakpoints } from "../lib/breakpoints.ts";
 import { flattenSpec } from "../lib/flatten.ts";
 import { pascalCase } from "../lib/pascal-case.ts";
 import { loadVocabulary } from "../lib/vocabulary.ts";
-import type { GeneratorContext, GeneratorReport } from "../registry.ts";
-import { registerGenerator } from "../registry.ts";
 import { Spec as SpecSchema } from "../schema.ts";
 import type { Spec } from "./gen-contract.ts";
 import { renderCssShim } from "./gen-vue-3/_shared/css-shim.ts";
